[LM-4530] event recordings and attachments (#4083)
added list of event_attachment objs in res, added updation of max_last_updated for sync_apis

+++ project/togther/models.py
@@ -2690,13 +2690,13 @@ class EventRecordingsAttachments(models.Model):
             'multimedia position'
         )
     )
-    width = models.FloatField(
+    width = models.IntegerField(
         null=True,
         help_text=_(
             'width of multimedia'
         )
     )
-    height = models.FloatField(
+    height = models.IntegerField(
         null=True, 
         blank=True,
         help_text=_(
